creative-project-module7/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
behave==1.2.6
blinker==1.6.2
bs4==0.0.1
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
flask-behind-proxy==0.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.0
get-chrome-driver==1.3.12
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
helium==3.2.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
packaging==23.1
pandas==2.0.2
parse==1.19.1
parse-type==0.6.2
Pillow==9.5.0
pluggy==1.2.0
pycodestyle==2.10.0
pymssql==2.2.7
pypng==0.20220715.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==3.141.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==3.8.6
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
